requirements.txt
pyperclip~=1.8.0
PyYAML~=5.3.1
PyQt5~=5.14.2
requests~=2.23.0
waitress~=1.4.3
Flask~=1.1.2